pyproject.toml
[tool.poetry]
name = "madr_novels"
version = "0.1.0"
description = ""
authors = ["amanmdest <amanmdest@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13, <4.0"
fastapi = {extras = ["standard"], version = "^0.115.0"}
alembic = "^1.13.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
pydantic-settings = "^2.6.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
python-multipart = "^0.0.17"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.6.7"
taskipy = "^1.13.0"
httpx = "^0.27.2"
factory-boy = "^3.3.1"
ignr = "^2.2"
freezegun = "^1.5.1"
testcontainers = "^4.9.2"
pytest-asyncio = "^1.0.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev madr_novels/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=madr_novels -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
